+++ logic/src/main/java/application/Repository/CompanyRepository.java
@@ -4,6 +4,6 @@ import application.models.Company;
 import org.springframework.data.repository.CrudRepository;
 
 
-public interface CompanyRepository extends CrudRepository<Company, Integer> {
+public interface CompanyRepository extends CrudRepository<Company, Long> {
     Company findById(long companyId);
 }
